backend/PySrc/requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
APScheduler==3.11.0
asgiref==3.8.1
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
boto3==1.35.90
botocore==1.35.90
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Faker==35.2.0
filelock==3.17.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
influxdb-client==1.48.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jwt==1.3.1
lxml==5.3.1
MarkupSafe==3.0.2
multidict==6.1.0
mysqlclient==2.2.7
newsapi-python==0.2.7
numpy==2.2.2
oauthlib==3.2.2
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
propcache==0.3.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
reactivex==4.0.4
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.1
setuptools==75.6.0
six==1.17.0
# snscrape @ git+https://github.com/JustAnotherArchivist/snscrape.git@614d4c2029a62d348ca56598f87c425966aaec66
soupsieve==2.6
SQLAlchemy==2.0.28
threadpoolctl==3.5.0
tweepy==4.15.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
update-checker==0.18.0
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
xgboost==2.1.4
yarl==1.18.3
